backend/requirements.txt
psycopg2-binary==2.9.9
asyncio==3.4.3 